requirements.txt
# Data handling
pandas==2.3.2
duckdb==1.3.2

# API backend
fastapi==0.116.1
uvicorn==0.35.0

# AI/LLM
openai==1.105.0
langchain==0.3.27
langgraph==0.6.6

# Data generation (for initial CSV creation)
faker==37.6.0

# Environment variables
python-dotenv==1.1.1

gradio==5.44.1
langchain-openai == 0.3.32